backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
email-validator
slowapi
pypdf
python-dotenv
google-genai
openai
boto3
httpx